Added code coverage generation with llvm tools
Using flags passed into clang to have the tests generate a profraw file
each. Those files are then merged into a profdata with llvm-profdata.
That file can then be used to show results of the coverage tracking with
llvm-cov.

These actions have been scripted into generate_coverage_report.py to
generate an html tree in the build folder as well as print a summary of
the coverage report when using `make coverage`.

Signed-off-by: Cora Weidner <cweidner3+git@gmail.com>